+++ documentation/website/.dir-template/src/docs/intro/change-log.md
@@ -22,6 +22,9 @@ Starred releases are most stable ones.
 - Add NutsDescriptor/NutsDescriptorBuilder::setLicenses
 - Add NutsDescriptor/NutsDescriptorBuilder::setDevelopers
 + Rename NutsConstants.IdProperties.DESKTOP_ENVIRONMENT -> NutsConstants.IdProperties.DESKTOP
++ Add    NutsConstants.IdProperties.DESKTOP_ENVIRONMENT -> NutsConstants.IdProperties.PROPERTIES
++ Rename NutsRepositoryDB::getRepositoryNameByURL -> NutsRepositoryDB::getRepositoryNameByLocation
++ Rename NutsRepositoryDB::getRepositoryURLByName -> NutsRepositoryDB::getRepositoryLocationByName
 
 ## nuts 0.8.3.1 (PUBLISHED VERSION)
 - ```2022/02/01 	nuts 0.8.3.1``` released nuts-runtime-0.8.3.1.jar